Refactor ShootingCSV to be a top level function
